Started moving UI components to UI folder
