+++ src/PhotosphereEditor.tsx
@@ -17,7 +17,7 @@ import {
 import { deleteStoredVFE, save } from "./FileOperations.ts";
 import { VisitedState } from "./HandleVisit.tsx";
 import PhotosphereViewer from "./PhotosphereViewer.tsx";
-import { alertMUI, confirmMUI } from "./StyledDialogWrapper.tsx";
+import { alertMUI, confirmMUI } from "./UI/StyledDialogWrapper.tsx";
 import {
   HotspotUpdate,
   convertRuntimeToStored,
