+++ src/PopOver.tsx
@@ -22,7 +22,7 @@ import Box from "@mui/material/Box";
 
 import { Asset, Hotspot2D, Hotspot3D, HotspotData } from "./DataStructures";
 import HotspotEditor, { HotspotIcon, NestedHotspotBox } from "./HotspotEditor";
-import { confirmMUI } from "./StyledDialogWrapper";
+import { confirmMUI } from "./UI/StyledDialogWrapper";
 import { HotspotUpdate } from "./VFEConversion";
 
 interface HotspotContentProps {
