+++ src/UI/StyledDialogWrapper.tsx
@@ -1,3 +1,9 @@
+// StyledDialogWrapper.tsx
+// This file wraps our functions in the StyledDialog.tsx
+// for react-specific purposes. This one exports the 
+// necessary functions our clients should use to make
+// the confirmation pop ups appear on the browser
+
 import { createConfirmation } from "react-confirm";
 
 import {
