+++ src/buttons/RemovePhotosphere.tsx
@@ -15,7 +15,7 @@ import { VFE } from "../DataStructures";
 import PhotosphereSelector, {
   PhotosphereSelectorProps,
 } from "../PhotosphereSelector";
-import { alertMUI } from "../StyledDialogWrapper";
+import { alertMUI } from "../UI/StyledDialogWrapper";
 
 interface RemovePhotosphereProps {
   onClose: () => void;
